pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "dtu_mlops_111"
version = "0.0.1"
description = "A short description of the project."
authors = [
  { name = "Your name (or your organization/company/team)", email = "your@email.com" },
]
license = { file = "LICENSE" }
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi==0.115.6",
    "kaggle==1.6.12",
    "matplotlib>=3.10.8",
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "torch==2.8.0",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "python-multipart>=0.0.9",
    "pillow>=10.0.0",
    "scikit-learn>=1.4.0",
    "dotenv>=0.9.9",
    "nnunetv2>=2.6.2",
    "dvc-gs>=3.0.2",
    "dvc>=3.66.1",
    "bentoml>=1.2",
    "types-pygments>=2.19.0.20251121",
    "types-defusedxml>=0.7.0.20250822",
    "types-tqdm>=4.67.0.20250809",
]

[dependency-groups]
dev = [
    "coverage==7.6.9",
    "invoke==2.2.0",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
    "mypy>=1.19.1",
    "httpx",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
lint.select = ["I"]
line-length = 120

[tool.coverage.run]
omit = ["tests/*"]

[[tool.mypy.overrides]]
module = ["dtu_mlops_111.*"]
follow_untyped_imports = true

[tool.pytest.ini_options]
# Add "src" to pythonpath so pytest can find the dtu_mlops_111 package
pythonpath = ["src"]
filterwarnings = [
    "ignore::DeprecationWarning:batchgenerators.*",
    "ignore::DeprecationWarning:importlib.*",
]
